[refactor] 기존 리덕스 패턴을 ducks패턴으로 바꿈
