--- sidebar/src/modules/actions/actionTypes.js
@@ -1,8 +1,0 @@
-export const INPUT_TITLE = 'createbucket/INPUT_TITLE';
-export const INPUT_DESC = 'createbucket/INPUT_DESC';
-export const ADD_DETAIL = 'createbucket/ADD_DETAIL';
-export const UPDATE_DETAILDUE = 'createbucket/UPDATE_DETAILDUE';
-export const REMOVE_DETAIL = 'createbucket/REMOVE_DETAIL';
-export const LOAD_PRESET = 'createbucket/LOAD_PRESET';
-export const CREATE_BUCKET = 'createbucket/CREATE_BUCKET';
-export const RESET_BUCKET = 'createbucket/RESET_BUCKET';
